Se creo un nuevo componente y se le agrego mas contenido al main
